refactor: move upload helpers to services
