+++ services/upload/__init__.py
@@ -1,9 +1,13 @@
 from .processing import UploadProcessingService
-from .ai import AISuggestionService
+from .ai import AISuggestionService, analyze_device_name_with_ai
 from .modal import ModalService
+from .helpers import get_trigger_id, save_ai_training_data
 
 __all__ = [
     "UploadProcessingService",
     "AISuggestionService",
     "ModalService",
+    "analyze_device_name_with_ai",
+    "get_trigger_id",
+    "save_ai_training_data",
 ]
